requirements.txt
Django==3.2.7
djangorestframework==3.12.4
python-dotenv==0.19.0
psycopg2-binary==2.8.6
django-filter==2.4.0